Release version 1.0.1
git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@70 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ samples/reflection/method.cpp
@@ -58,7 +58,7 @@ public:
 	}
 	
 	int calcObj(const TestObject & other) {
-		return other.width + other.name.length();
+		return static_cast<int>(other.width + other.name.length());
 	}
 
 	int sum(const cpgf::GMetaVariadicParam * params) const {
